rename dns.c -> main.c, request.c -> dns.c and request.h -> dns.h
